refactoring minor things
